Add Quaternion and MD5 implementations; refactor includes and create UI components
- Introduced Quaternion class for quaternion operations including multiplication, conversion to matrix, and spherical linear interpolation (Slerp).
- Implemented a minimal MD5 hashing class with methods for initialization, updating, finalizing, and generating hexadecimal digest.
- Refactored include paths in Camera, FrameBuffer, and other engine files to use the new commons directory structure.
- Created SCButton class for UI button functionality, including click handling and appearance management.
- Added SCZone class for managing interactive zones in the UI, including drawing and action handling.
- Updated various files to ensure proper integration of new components and maintain consistent coding standards.
