+++ src/commons/ByteStream.cpp
