+++ src/commons/ByteStream.h
