+++ src/commons/IffLexer.cpp
