+++ src/commons/LZBuffer.cpp
