+++ src/commons/Quaternion.cpp
