+++ src/engine/IActivity.cpp
@@ -6,7 +6,7 @@
 //  Copyright (c) 2014 Fabien Sanglard. All rights reserved.
 //
 
-#include "precomp.h"
+#include "IActivity.h"
 
 char *strtoupper(char *dest, const char *src) {
     char *result = dest;
