+++ src/sc/GameEngine.cpp
@@ -9,7 +9,7 @@
 #include "imgui.h"
 #include "imgui_impl_opengl2.h"
 #include "imgui_impl_sdl2.h"
-#include "precomp.h"
+#include "GameEngine.h"
 
 #define FLOPPY 0
