+++ src/sc/GameEngine.h
@@ -8,11 +8,15 @@
 
 #ifndef __libRealSpace__Game__
 #define __libRealSpace__Game__
-#include "precomp.h"
+
 #define HEIGHT 800
 #define WIDTH 1200
 #define FULLSCREEN 0
+#include <stack>
+#include <vector>
+#include <string>
 
+class IActivity;
 class GameEngine{
     
 public:
